bugfix: rename Invitations -> Invites to fix travis issue recognizing component
