+++ client/components/AllOrgs.js
@@ -1,7 +1,7 @@
 import React, {useState, useContext, useEffect} from 'react'
 import Modal from './Modal'
 import LeaveOrgModal from './LeaveOrgModal'
-import Invitations from './Invitations'
+import Invites from './Invites'
 import {Link} from 'react-router-dom'
 import {AuthContext} from '../context/authContext'
 import InvitationProvider, {
